Test fix

+++ database/migrations/2019_07_10_100001_create_vkontakte_users_table.php
@@ -14,7 +14,7 @@ class CreateVkontakteUsersTable extends Migration
     public function up()
     {
         Schema::create('vkontakte_users', function (Blueprint $table) {
-            $table->integer('id')->primary();
+            $table->bigInteger('id')->primary();
 
             $table->string('first_name');
             $table->string('last_name');
